Implemented Avatar and Alert Dialogue from Radix-UI from the application.

+++ src/components/Tokens.tsx
@@ -1,5 +1,6 @@
 import styles from '@/styles/Tokens.module.css';
 import React from "react";
+import * as AlertDialog from '@radix-ui/react-alert-dialog';
 import { useState, useEffect } from "react";
 import {
   useActivateTokensMutation,
@@ -147,17 +148,52 @@ const Tokens = (props: any) => {
                         )}
                       </td>
                       <td>
-                        {token.status !== "Deleted" && (
-                          <button
-                            type="button"
-                            className="btn"
-                            onClick={() => {
-                              handleDeleteToken(token.id);
-                            }}
-                          >
-                            Delete
-                          </button>
-                        )}
+                        <AlertDialog.Root>
+                          <AlertDialog.Trigger asChild>
+                            {token.status !== "Deleted" && (
+                              <button
+                                type="button"
+                                className="btn"
+                              >
+                                Delete
+                              </button>
+                            )}
+                          </AlertDialog.Trigger>
+                          <AlertDialog.Portal>
+                            <AlertDialog.Overlay className="AlertDialogOverlay" />
+                            <AlertDialog.Content className="AlertDialogContent">
+                              <AlertDialog.Title className="AlertDialogTitle">
+                                Are you sure?
+                              </AlertDialog.Title>
+                              <AlertDialog.Description className="AlertDialogDescription">
+                                This action cannot be undone. This will
+                                delete your token permanently.
+                              </AlertDialog.Description>
+                              <div
+                                style={{
+                                  display: "flex",
+                                  gap: 25,
+                                  justifyContent: "flex-end",
+                                }}
+                              >
+                                <AlertDialog.Cancel asChild>
+                                  <button className="Button mauve">
+                                    Cancel
+                                  </button>
+                                </AlertDialog.Cancel>
+                                <AlertDialog.Action asChild>
+                                  <button className="Button red" 
+                                    onClick={() => {
+                                      handleDeleteToken(token.id);
+                                    }}
+                                  >
+                                    Yes, Delete Token
+                                  </button>
+                                </AlertDialog.Action>
+                              </div>
+                            </AlertDialog.Content>
+                          </AlertDialog.Portal>
+                        </AlertDialog.Root>
                       </td>
                     </tr>
                   );
